backend/requirements.txt
# Web framework & server
fastapi
uvicorn[standard]

# HTTP client
httpx

# Data validation & settings
pydantic
pydantic-settings
python-dotenv

# JSON & logging
orjson
loguru

# LLM providers 
openai
anthropic
google-generativeai

# Reliability & dates
tenacity
python-dateutil

# Database, auth & crypto
SQLAlchemy>=2.0
aiosqlite
passlib[bcrypt]
python-jose[cryptography]

# Quality
ruff
black
